feat(STK-199): create constants and types for FA usage
