+++ packages/app/constants/events.ts
@@ -1,0 +1,7 @@
+export type EventName = string;
+
+export const EVENTS = {
+  CLICK: {
+    STRATEGY_CARD: "click/strategy-card",
+  },
+};
